update get test
