+++ packages/core/test/functions/counterYe.js
@@ -1,0 +1,3 @@
+module.exports = (_payload, update) => {
+  update({ yeye: 'yeye' })
+}
