+++ packages/core/test/functions/errorFunction.js
@@ -1,0 +1,5 @@
+module.exports = async () => {
+  const wawa = [1, 2]
+  // @ts-ignore
+  return wawa[3].yeye
+}
