+++ packages/core/test/functions/errorTimer.js
@@ -1,0 +1,6 @@
+module.exports = (payload, update) => {
+  setInterval(() => {
+    throw new Error('lol')
+  }, 10)
+  update('yes')
+}
