+++ packages/core/test/functions/throwingFunctions.js
@@ -1,0 +1,3 @@
+module.exports = async () => {
+  throw new Error('This is error message')
+}
